requirements.txt
fastapi==0.95.0
uvicorn==0.18.2
pydantic==1.10.2
python-dotenv==0.21.0
requests==2.28.1
pymongo==4.3.3
pydantic[email]==2.7.1
